2004-08-11	blytkerchan 	* array.c (array_put): use compare_and_exchange_ptr in stead of 	obsolete compare_and_exchange 	* binomial_tree.c (binomial_tree_get): ditto 	(binomial_tree_set_root): ditto 	(binomial_tree_node_set_left): ditto 	(binomial_tree_node_set_right): ditto 	(binomial_tree_node_set_value): ditto 	* list.c (list_find): ditto 	(list_delete): ditto 	* map.c (map_remove): ditto 	* queue.c (queue_enq): ditto 	(queue_deq): ditto 	* stack.c (free_stack): ditto 	(stack_pop): ditto 	(stack_push): ditto 	* heap.c (heap_select_node): use uint32_t for exp 	use compare_and_exchange_int in stead of obsolete compare_and_exchange 	(heap_fix): ditto (both) 	(heap_add): use uint32_t for o_n, n and exp 	use compare_and_exchange_int in stead of obsolete compare_and_exchange 	* heap.h: include stdint.h 	[heap_node_t]: use uint32_t for flag in stead of int 	[heap_t]: ditto for N and flag

+++ heap.h
@@ -34,6 +34,7 @@
 #ifndef _LIBCONTAIN_HEAP_H
 #define _LIBCONTAIN_HEAP_H
 
+#include <stdint.h>
 #include "libcontain_config.h"
 
 #define HEAP_DEFAULT_SIZE 	100
@@ -44,16 +45,16 @@
 
 typedef struct {
 	void * val;
-	int flag;
+	uint32_t flag;
 } heap_node_t;
 
 typedef struct 
 {
 	binomial_tree_t * tree;
 	libcontain_cmp_func_t heap_val_cmp_func;
-	size_t N;
+	uint32_t N;
 	unsigned int workers;
-	int flag;
+	uint32_t flag;
 } heap_t;
 
 heap_t * heap_new(libcontain_cmp_func_t heap_val_cmp_func);
